% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/say.R
\name{say}
\alias{say}
\title{Say hi}
\usage{
say(name = "Andrea")
}
\arguments{
\item{name}{the name to say hi to (character)}
}
\value{
Nothing, side-effect of a message
}
\description{
Say hi
}
\examples{
say()
say("Andrea")
}
