% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testytest-package.R
\docType{package}
\name{testytest-package}
\alias{testytest}
\alias{testytest-package}
\title{testytest: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/maelle/testytest}
  \item Report bugs at \url{https://github.com/maelle/testytest/issues}
}

}
\author{
\strong{Maintainer}: Maëlle Salmon \email{msmaellesalmon@gmail.com} (\href{https://orcid.org/0000-0002-2815-0399}{ORCID})

}
\keyword{internal}
